redireccionar a mis cursos despues de login y ocultar la dashboard a mis cursos (esa estructura debe tener) eliminar asistencias y usar ese codigo en mis cursos, limpiar codigo sucio, modificar dashboard y eliminar importacion de asistencias en el main
