+++ frontend/src/components/Dashboard.jsx
@@ -7,7 +7,6 @@ import SchoolIcon from '@mui/icons-material/School';
 import { DashboardLayout } from '@toolpad/core/DashboardLayout';
 import { useNavigate, Outlet } from "react-router-dom";
 import { logout } from '@services/auth.service.js';
-import { Typography } from '@mui/material';
 
 
 const NAVIGATION = [
@@ -17,7 +16,7 @@ const NAVIGATION = [
     icon: <SchoolIcon />,
   },
   {
-    segment: 'Asistencias',
+    segment: 'Cursos',
     title: 'Cursos',
     icon: <SchoolIcon />,
   },
@@ -95,6 +94,9 @@ function DashboardLayoutAccount() {
     };
   }, [navigate]);
 
+  const hideNavigation = location.pathname === '/Cursos';
+
+
   return (
     <ThemeProvider theme={customTheme}>
       <Box sx={{ minHeight: '100vh', background: (theme) => theme.palette.background.default }}>
@@ -108,7 +110,7 @@ function DashboardLayoutAccount() {
             title: "SmeBook", // Deja esto como un string
           }}
         >
-          <DashboardLayout>
+          <DashboardLayout hideNavigation={hideNavigation}>
             {/* Renderiza el contenido directamente si el usuario está autenticado */}
             <Outlet />
           </DashboardLayout>
