+++ frontend/src/main.jsx
@@ -12,7 +12,6 @@ import Notas from '@pages/Notas';
 import Labs from '@pages/Labs'; // Importa el nuevo componente Labs
 import VerNotas from '@pages/VerNotas';
 import '@styles/styles.css';
-import Asistencias from '@pages/Asistencias';
 import VerAsistencias from '@pages/VerAsistencias';
 import RegistrarAsistencias from '@pages/RegistrarAsistencias';
 
@@ -27,7 +26,6 @@ const router = createBrowserRouter([
       { path: 'anotaciones', element: <Anotaciones /> },
       { path: 'notas', element: <Notas /> },
       { path: 'labs', element: <Labs /> }, // Agrega la nueva ruta para Labs
-      { path: 'asistencias', element: <Asistencias /> },
       { path: 'VerAsistencias/:id_curso', element: <VerAsistencias /> }, 
       { path: 'RegistrarAsistencias/:id_curso', element: <RegistrarAsistencias /> }, 
       {
